updated permissions; rsolved bugs

+++ src/main/java/com/batherphilippa/pin_it_app_be/controller/ProjectController.java
@@ -2,7 +2,10 @@ package com.batherphilippa.pin_it_app_be.controller;
 
 import com.batherphilippa.pin_it_app_be.dto.*;
 import com.batherphilippa.pin_it_app_be.exceptions.ProjectNotFoundException;
+import com.batherphilippa.pin_it_app_be.exceptions.UserNotAuthorisedException;
 import com.batherphilippa.pin_it_app_be.exceptions.UserNotFoundException;
+import com.batherphilippa.pin_it_app_be.model.Permissions;
+import com.batherphilippa.pin_it_app_be.model.ProjectUser;
 import com.batherphilippa.pin_it_app_be.model.User;
 import com.batherphilippa.pin_it_app_be.service.ProjectService;
 import com.batherphilippa.pin_it_app_be.service.UserService;
@@ -54,32 +57,25 @@ public class ProjectController {
         logger.info("ProjectController: saveProject");
         return new ResponseEntity<>(projectDTOOut, HttpStatus.CREATED);
     }
-/*
-    // TODO
-    @PostMapping("/projects/users/{userId}")
-    public ResponseEntity<ProjectAndPermissionsDTOOut> addUserToExistingProject(@PathVariable String email, @Valid @RequestBody ProjectPermissionsDTOIn projectPermissionsDTOIn) throws UserNotFoundException {
-        return null;
-    }*/
-
-    @PutMapping("/project/{projectId}")
-    public ResponseEntity<ProjectAndPermissionsDTOOut> updateProjectById(@PathVariable long projectId, @Valid @RequestBody ProjectDTOIn projectDTOIn) throws ProjectNotFoundException {
+    @PutMapping("/project/{projectId}/user/{userId}")
+    public ResponseEntity<ProjectAndPermissionsDTOOut> updateProjectById(@PathVariable long projectId, @PathVariable long userId, @Valid @RequestBody ProjectDTOIn projectDTOIn) throws UserNotAuthorisedException, ProjectNotFoundException {
+        ProjectUser projectUser = projectService.validatePermissions(projectId, userId);
+        if(projectUser.getPermissions() == Permissions.VIEWER) {
+            throw new UserNotAuthorisedException(userId);
+        }
         projectService.updateProjectById(projectId, String.valueOf(projectDTOIn.getDeadline()), projectDTOIn.getDescription(), projectDTOIn.getTitle());
         logger.info("ProjectController: updateProjectById");
         return ResponseEntity.noContent().build();
     }
 
-    @DeleteMapping("/project/{projectId}")
-    public ResponseEntity<Void> deleteProjectById(@PathVariable long projectId) throws ProjectNotFoundException {
+    @DeleteMapping("/project/{projectId}/user/{userId}")
+    public ResponseEntity<Void> deleteProjectById(@PathVariable long projectId, @PathVariable long userId) throws UserNotAuthorisedException, ProjectNotFoundException {
         logger.info("ProjectController: deleteProjectById");
+        ProjectUser projectUser = projectService.validatePermissions(projectId, userId);
+        if(projectUser.getPermissions() == Permissions.VIEWER || projectUser.getPermissions() == Permissions.EDITOR_RW) {
+            throw new UserNotAuthorisedException(userId);
+        }
         projectService.deleteProjectById(projectId);
         return ResponseEntity.noContent().build();
     }
-/*
-    @GetMapping("/projects/owned-projects/{userId}")
-    public ResponseEntity<List<OwnerProjectGuestsDTOOut>> getOwnedProjectsAndGuests(@PathVariable long userId) throws UserNotFoundException {
-        User user = userService.findById(userId);
-        List<OwnerProjectGuestsDTOOut> ownedGuestProjectsSet = projectService.getGuestsOnOwnerProjects(user.getEmail(), user.getId());
-        return new ResponseEntity<>(ownedGuestProjectsSet, HttpStatus.OK);
-    }*/
-
 }
